[go][concurrencyrace] Let everyone has a chance to get good food for each race
Summary:
Let everyone has a chance to get good food for each race by getting a random
sleep before getting his/her food.

Remove the unnecessary "sync.Once" for getting the food.

Adjust the time duration, the text, and the number of people to make the
program a little bit more interesting.

Make sure history is done before the main program exits, so each line of code
in the program is guaranteed to be run (previously main exits before the
goroutine that prints out history can print that its done).

Test Plan:
Run the following a few times and see that everyone has a chance to get an
apple.
go run concurrencyrace.go