[android][PhotoGallery] Created PhotoGallery and finished Ch 23. HTTP & Background Tasks
I implemented a way to hide api key while open sourcing my code by storing the
api key in the device in the first run, so subsequent runs can just read the
api key from the file.

I implemented the third challenge that uses dynamic columns.

I did not implement the paging challenge because it is a bit time consuming.