[loc8r] Add api's for reading a location or a review
It's part of Ch.6 of the book "Getting MEAN with Mongo, Express, Angular, and
Node".

I found that the subdocument needs to have the field "_id" instead of
"id" in order for a method call like location.reviews.id(<id value
here>) to work, so I modified the data in db_setup.