[loc8r] Implement locationsListByDistance
It's part of Ch.6 of the book "Getting MEAN with Mongo, Express,
Angular, and Node".

Implement locationsListByDistance by using geoNear method of mongoose
model. Notice that the implementation does not use the
getDistanceFromRads and getRadsFromDistance methods as shown in the book
because the book probably has some inconsistent information. Since we
are parsing a GeoJSON to geoNear, the parameters should be in meters,
same as the distance from the output.

Fix some error handling code.