[loc8r] Implement locationsCreate, locationsUpdateOne, reviewsCreate
It's part of Ch.6 of the book "Getting MEAN with Mongo, Express,
Angular, and Node". For some reason, openingTimes cannot be persisted. I
suspect that I need to update them the way that I create a review,
menaing it needs a separate step to update them instead of saving them
along with the rest of a location.

Add sample_http_calls folder to contain sample http calls to the api for
testing purpose.

Use the "underscore" library to make the code cleaner.